Overhauling error module.
Making a base class with a heavily overloaded constructor method that
populates the sole attribute to hold the error report. The base class
requires a name argument, which is hard-coded by child classes.

The previous set up was having issues with the error messages being
binary gibberish. This overhaul seems to fix this issue.