Overhauling pattern folding in BiallelicData.
Folding all patterns, not just mirrored patterns. In other words,
calling 'fold_patterns' will ensure that every pattern is the "mirror"
version in which the number of red alleles is less frequent (if the
number of red alleles is equal in the original and mirrored version, the
original is retained).

Due to this change, the boolean method 'patterns_are_folded' changed
meaning and is no longer the opposite of 'has_mirrored_patterns'.  It
now returns false if a pattern and its mirror are present, OR any
pattern has more red alleles than its corresponding (but not present)
mirror pattern.

All tests are updated and passing.