Updating and adding tests of parameter classes.
This brings tests up-to-date with changes in last commit. All
passing.

I forgot to mention in last commit message, that the default value for
all real variables/parameters in now 'NaN', and there is now a new
method 'initialize_value' that will set the value from the prior if the
current value is 'NaN'.