Updating tree classes.
Deleting 'set_child_coalescence_rate_parameter' method from
ComparisonPopulationTree, because it is not used and creates an
unnecessary risk of getting mixed priors on the same tree.

Adding an assert to 'set_root_height_paramter' to ensure the new
parameter's prior points to 'this->node_height_prior_').

All tests passing.