Overhauling creating instances from settings.
Removing most of 'get_instance' method from settings classes; instead,
adding constructors for PositiveRealParameter and
ComparisonPopulationTree that take their respective settings classes as
an argument. This clears a lot of imports in the settings module, and is
a bit cleaner. The only 'get_instance' method remaining in settings is
for ContinuousDistributionsettings to return a
ContinuousProbabilityDistribution pointer. Because this method returns a
base class pointer, pointing to one of the derived probability classes,
it cannot be ported to a constructor.

All test passing, but I removed the operator and collection modules from
building/testing, so these are still broken.