Overhauling of priors, parameters, operators.
Removing 'v' rate parameter completely. The 'v' rate is simply a
deterministic function of the 'u' rate, and thus maintaining it for the
sole purpose of allowing a prior to be placed on it is silly.  Updating
PopulationTree, settings, and associated tests accordingly.

Adding MutationRateScaler and ComparisonHeightMover operators.  Making
MutationRateScaler the operator of choice (the only choice in the
settings) for updating the 'u' rate. The MutationRateMover is not quite
returning the prior, and I think this is due to the updating occuring on
a different scale (the frequency of red allele) than the 'u' parameter.
Updating settings and OperatorSchedule constructor accordingly.

Adding more operator tests. All tests are passing except those for
RootCoalescenceRateScaler and ChildCoalescenceRateScaler; they are not
returning the prior without data.