Fixing bug in DirichletProcessGibbsSampler.
Fixing bug in DirichletProcessGibbsSampler operator. Previously, I was
using the vector 'ln_category_likelihoods' to get the tree likelihood
when updating its root height. But, this was not the tree log
likelihood! It is the tree log likelihood multiplied by the DPP prior of
that category. This commit fixes this problem by adding a second vector
which only stores the tree log likelihood in each proposed category.

The DPP prior sampling test is now passing! Woohoo!