Fixing concentration bug.
Updating ComparisonPopulationTreeCollection methods. Updating
'store_state' and 'restore_state' so that they now store and restore the
concentration parameter (this was missing).  Updating
'compute_log_likelihood_and_prior' to include log prior density of the
concentration parameter in the log prior calculation (which was
missing).

This changes fixed the issue of sampling from the prior of concentration
parameter.

All tests passing!!