More work on rjMCMC.
Overhauling how ReversibleJumpSampler stores/restores the model state.
This is now handled by ComparisonPopulationTreeCollection.  This appears
to have fixed the bug in the original rjMCMC algorithm (now
ReversibleJumpSampler.propose_jump_to_gap). Both tests of
ReversibleJumpSampler are passing, but more are needed to make sure this
fix is general.

The ReversibleJumpSampler.propose_jump_to_prior algorithm is still
broken (if 'ReversibleJumpSampler.propose' is changed to use this
algorithm, the tests fail).  It seems "close," but the merge move seems
to be rejected too often.

Moving the storing/restoring of node heights from
ComparisonPopulationTree to ComparisonPopulationTreeCollection.  This
was not a problem, but this approach is cleaner (not calling
store/restore multiple times for shared parameters).

All tests are currently passing!