Overhauling treatment of constant and missing data.
By default, the main CLI of ecoevolity is more strict about
constant sites and populations with missing data.

If 'constant_sites_removed = True' is specified for an alignment, but a
constant site is found, an EcoevolityConstantSitesError is thrown.
Adding a '--relax-constant-sites' option to CLI to relax this, in which
case, the constant sites are ignored and a warning is reported.

If a data column is encountered for which at least one population has no
data, an EcoevolityMissingDataError is thrown. Adding a
'--relax-missing-sites' optiono to CLI to relax this, in which case, the
missing data columns are ignored and a warning is reported.

Updating/adding test data sets, and adding tests of this new behavior.

All test passing.