Adding more vetting of user settings.
Throwing an error if the population sizes are fixed and/or constrained
for singleton comparison (can't estimate pop expansion/contraction
timing if the sizes are constrained!).

Throwing an error if population labels are duplicated (i.e., if the same
population label is used in multiple alignments).

Tests of these errors have been added, and are passing.