Finishing overhaul of mutation multiplier.
Finishing change over from the use of 'rate_multiplier' to
'mutation_rate' throughout the code.

Also, finishing the addition of the ploidy parameter, which can now be
specified in the configuration file. The 'population_size' parameters
now represent Ne (or Ne * mutation rate) rather than 4Ne, because the
ploidy constant is multiplied just prior to calculating the likelihood
(or simulating a gene tree).

All tests are passing.