Overhauling parsing of nexus underscores.
Just going with the nexus standard of converting unquoted underscores to
spaces in sequence labels, rather than the hacky trick of converting '_'
population name delimiters to ' ' behind the scenes.

Updating code throughout package for this new behavior, and adding a few
tests to make sure the behavior is working.

All tests passing.