Consolidating time and pop size scaling.
Adding 'get_node_theta' and 'get_node_length_in_subs_per_site' methods
to PopulationTree. This takes care of the simple calculations of making
sure population size and time are in the right scale before either
calculating the likelihood or simulating a gene tree. Previously, these
simple calculations were done on-the-fly in multiple places in the code.
This makes sure the same scaling is being done everywhere.

All tests passing.