In middle of overhaul of u/v --> frequencies
Changing the use of u/v rates over to dealing with the stationary
frequencies of the 0/1 states. u/v/freq_0/freq_0 are all deterministic
functions of one another (you only need one to get the other three).
Previously, I was doing everything in terms of u, but u (and v) has an
odd sample space (0.5-infinity), which requires the use offset priors.
Changing everything over to freq_1, which has the sample space 0-1.
This will make it easier to specify priors.

I'm currently in the middle of this transition, so everything is likely
broken.