Moving likelihood calc out of tree module.
Moving likelihood calcs into separate module. This should make
multithreading across site patterns easier. Currently in the middle of
this move, so everything is broken.

Also, working on adding deep copying to PopulationNode. This will be
necessary for the multithreading (so threads are not trying to update
the same BiallelicProbabilityMatrix instances during the likelihood
comps).