Fixing bug when constant site counts provided.
Fixing bug in likelihood comp when constant site counts are provided to
PopulationTree. Accidently removed std::log() of constant pattern
likelihoods during the great likelihood comp migration. Adding
it back in.

All tests are passing again! Caveat: the multithreading bit of the
likelihood module is currently commented out, because it is not
compiling.