Updating threaded get_log_likelihood.
Removing std::cref() wrapper around BiallelicData const ref in the async
call to get_log_likelihood_for_pattern_range call. This seems to have
been redundant (there seems to be no differnce with or without it).

Also, I tried to make copies of the BiallelicData dataset for each
thread to see if accessing the same const ref instance was slowing down
the threads. Hoever, there was no difference in performance, so I've
taken this copying back out.

All tests are passing, and there are no memory leaks with or without
multithreading (checked with valgrind).