Updating mcmc logic for RJ move.
Only doing a single move when te RJ operator is drawn during MCMC.  I
was doing multiple RJ moves equal to the number of trees, however, this
might slow things down more than it's worth as the number of trees
increases. If mixing is an issue, the weight of the move can simply be
increased.