Fixing bug in simcoevolity.
Restoring the original comparison paths in the CollectionSettings
instance  after each simulation. During each simulation, the comparison
paths in the CollectionSetting instance need to be updated to match the
simulated alignments. However, swapping these paths requires the
settings instance to have the comparison paths from the original config.
So, after each replicate the original paths are now being restored (so
that the swap in the next iteration will work).