Adding alternative proposal for CollectionScaler.
Tried an element-wise proposal of all node heights and pop sizes for
CollectionScaler. This sampled from the expected prior when there were
no data. However, it did not improve the mixing of sampling node heights
and pop sizes when no constant sites are present. The new proposal
method is included, but commented out.