Adding tests of CollectionScaler operator.
All tests passing. CollectionScaler is a multivariate proposal that
scales all of the node height and pop size parameters. Alone, it cannot
sample the prior when there are no data. However, when coupled with the
univariate operators of node heights and pop sizes, the expected prior
distribution is sampled. This greatly improves mixing when conditioning
on no constant characters. Based on prelim analyses, affects on mixing
are minimal (maybe slightly better) when constant characters are present
(the mixing was already good without it).