Working on ReversibleJumpWindowOperator.
Fixing some bugs in the operator. Also, the operator now has an instance
of a ComparisonNodeMover as a member. This is used primarily to control
the autotuning of the width of the window for the rj move.  But, it also
proposes moves to all the node heights before a jump move is tried.

It seems to be close. It's passing prior sampling tests with 2 and 3
pairs, but not quite for 6-pairs. It's excruciatingly close to sampling
the prior for 6 pairs, but it is consistently slightly off.