Overhauling operator module.
I've been a bad developer. I established a branch with the intent of
committing fine-grained changes as I implemented a new reversible-jump
operator that was tunable. What ended up happening is a full
re-factoring of the entire operator module (Operator hierarchy of
classes).

I am adding a tunable RJ operator, but it is not being used. It works,
but is super sensitive to the size of the window, and will often fail to
sample the target (prior) distribution even when trying to auto-tune the
window size.

The re-factoring includes a new hierarchy 'OperatorInterface's, which is
now the interface to all MCMC moves. The base classes are templated to
have any descendant of 'Operator' as a member, and wrap up all of the
propose/accept/reject logic. The derived OperatorInterface classes
explicitly declare their 'Operator' member, and a 'propose' method to
use it. Client code now should call the 'operate' method of derived
OperatorInterface instances, which wraps up all the propose->calc
likelihood->accept/reject->restore logic.

All tests passing.