Overhauling multivariate operators.
Overhauling multivariate operators that used to be called
'CollectionScaler's. They are now derived from TimeOperatorInterface,
rather than CollectionOperatorInterface so that each call to 'propose'
only scales the time and pop sizes for associated to a single node
height (time) parameter. The 'perform_collection_move' method then walks
through and calls 'propose' on all of the node heights in the current
model state.

Also, updating ComparisonPopulationTreeCollection and base
'OperatorInterface' classes. Adding public methods to
ComparisonPopulationTreeCollection and using them in the base
'OperatorInterface' classes so that most operators no longer need to be
friends of the ComparisonPopulationTreeCollection class.

Adding lots of tests. All tests passing.