Generalizing the uniform model prior.
Adding a 'split_weight' parameter to the uniform model prior that allows
split/lumped models to be preferred. All divergence models within a
class (i.e., with the same number of divergence events) still have
uniform prior probability, but a weight can be assigned to favor classes
of models with more or fewer events.

Adding lots of tests; all passing.