Updating DirichletProcessGibbsSampler.
Making DirichletProcessGibbsSampler follow Radford Neal's (2000)
algorithm 8 exactly (barring any bugs, of course). The change entails
how to propose a Gibbs move to a tree in a singleton height category.
Previously, I was simply removing this category from consideration and
considering the tree goes in 'k-' other existing categories and 'm'
auxillary categories. Now, I consider the tree goes in 'k-' other
existing categores, it stays in its current singleton category, or 'm-1'
auxillary categories. Essentially, the current singleton category is
treated like one of the auxillary categories.

This will increase efficiency slightly (no likelihood calc for
considering staying in singleton category), and I'm hoping it will help
the poor behavior of the DPP model when the likelihood is not 1 (i.e.,
when the data are not ignored). I suspect it will not change it's
behavior, but it is worth a shot.

Tests to confirm that the DirichletProcessGibbsSampler is still sampling
from the prior when the data are ignored are passing.