Reworking pop size multiplier framework.
DirichletPopulationTree is now much more similar to PopulationTree;
there are no explicite multiplier parameters or mean population size
parameter. Now, the multipliers and mean pop size are only considered
when calculating the prior probablility of a tree, and when performing
MCMC moves on the tree. This set up makes it easier to implement
operators that help traverse the correlations among parameters.

Adding lots of new tests. All tests are passing.