Updating op schedule and multivariate ops
Updating the behavior of the sweep of calls made to univariate operators
after a multivariate time operator makes a move. If the multivariate
move is tree-specific, it will now only call the univariate moves for
that tree (rather than those for all trees).

This required new 'get_tree_operators' and 'get_time_operators' methods
for OperatorSchedule and BaseComparisonPopulationTreeCollection that
take the tree_index as an argument.