Updating tests.
Updating tests to reduce chance of some failures due to Monte Carlo.
Also, trying to reduce the number of iterations run for a lot of the
longer running tests (to reduce test-suite run time). I suspect some of
these shorter runs will trigger more failures due to MC error, but
hopefully this will reign in the the overall runtime of the test suite a
bit!