Adding more math_util functions for Pitman-Yor processes.
Adding functions to get the concentration parameter (or scale parameter
on a gamma prior on the concentration parameter) that corresponds to a
specified expected number of categories and discount parameter.

Also, updating get_dpp_concentration to avoid implicit casting of double
-> unsigned int when calling get_dpp_expected_number_of_categories.

Adding tests of new functions. All tests of the math_util module are
passing.