Adding warning message for 'uniform' model prior.
Analyzing simulated data has revealed some buggy behavior of the
reverisble-jump MCMC operator that is used for the 'uniform' model
prior. The Dirichlet-process prior is working great, and is the default,
so adding a very obnoxious warning message if the 'uniform' model prior
is specified.

All tests are currently passing.

Some notes on the buggy behavior for perpetuity:

When simulated datasets with 3-pairs are analyzed under very similar
Dirichlet-process and uniform model priors, I am seeing very different
results. Neither seem biased (true vs estimated pop size and div time
estimates line up nicely along the identity line). However, I get *much*
larger div time credibility intervals under the 'uniform' model, even
when the pop sizes are fixed!

Also, estimating the number of div times and div model is not as good
under the uniform model. Under the DPP model the number of div events is
correctly estimated 97+% of the time, whereas the uniform model is
correct ~ 84% time and tends to under estimate consistently.  This could
be a symptom of the really broad CIs on div times.

This seems to be a problem with the ReversibleJumpSampler. I am not sure
if this is a theoretical problem or a bug yet, but it is definitely not
a mixing issue (Monte Carlo error); everything is mixing beautifully,
and really long runs yield the same pattern.

I did some experiments with 3-pair datasets simulated and analyzed under
a uniform model prior and fixed population sizes (The only parameters
are the div times, and the only operators are the ReversibleJumpSampler
and EventTimeScaler). Here are my observations:

The scale parameter of the EventTimeScaler owned by the
ReversibleJumpSampler is getting autotuned to make *MUCH* larger moves
than the 'global' EventTimeScaler of the OperatorSchedule (the scale is
multiple orders of magnitude larger!).
The EventTimeScaler owned by the ReversibleJumpSampler updates the div
times after every RJ move.

I did an experiment where I removed the EventTimeScaler owned by
ReversibleJumpSampler; instead the 'global' EventTimeScaler gets
called after every RJ move. This caused the scale parameter of
the 'global' EventTimeScaler to be autotuned to intermediate values,
and the CIs on divergence times were reduced (but still much larger
than they should be). The correct div model was also sampled more.

Next, I reduced the number of times the RJ move was performed each time
the ReversibleJumpSampler was called; once instead of three times (the
number of pairs). The 'global' EventTimeScaler was still called after
each RJ (still no EventTimeScaler owned by the ReversibleJumpSampler).
This caused the autotuned scale values and CI intervals to be smaller
(but still much larger than under a DPP model). Again, the
correct model was sampled more.

Something seems very weird about the div times after RJ moves to cause
the EventTimeScaler to make such massive moves.  What's amazing is that
most of the time the RJ moves are rejected (~ 10% acceptance rate,
sometimes lower).

Also, the ReversibleJumpSampler samples from the joint prior just fine
when the data are ignored. So, the problem has something to do with the
influence of the likelihood, but doesn't cause any obvious bias (just
too much uncertainty).