Updating behavior of simcoevolity.
Udpating output behavior when the '--parameters-only' option is used.
When this option is specified, all output goes to stderr/stdout,
rather than being written directly to files.

Also, all info (with or without '--parameters-only' option) is now
written to stderr rather than stdout. That allows the parameter values
to be written to stdout (and redirected) when the '--parameters-only'
option is used.