Adding option to handle triallelic sites.
Adding option to handle sites with more than two character states. The
option allows such sites to be recoded as binary (first character found
is 0, all others are 1) and retained (rather than throwing an error and
exiting).

The recoding is handled in the BiallelicData class, and the option to
keep triallelic sites (and not throw an error) is handled by
constructors of population tree classes, which is called by the
collection classes.

The option is now available in the ecoevolity and simcoevolity
executables.

Adding/modifying associated tests and test alignments.

All tests of new behavior are passing; whole test suite still needs to
be run.