Updating simcoevolity.
The genotypes of output simulated alignments are always haploid, but the
output config files were not being updated to 'genotypes_are_diploid =
false'. So, simulating from a config that specified diploid data
produced configs that would fail to run.

Adding a method to collection settings classes to update all the
'genotypes_are_diploid' settings, and using this option in
simcoevolity_main.