New correction for constant patterns is complete.
Updating last test to be consistant with the new behavior of correcting
the likelihood for missing constant patterns.

Previously, the correction term that likelihoods were divided by
was:

    $ (1.0 - r_m - g_m)^n $

Where $n$ is the number of sites, and $r_m$ is the probability of 'all
red' allele for the maximum number of sampled alleles for all
populations, and $g_m$ is the probability of 'all green' alleles for the
maximun number of sampled alleles for all populations.

Now, we use:

    \[
    (1.0 - r_1 - g_1)^n_1 \times
    (1.0 - r_2 - g_2)^n_2 \times
    \ldots
    \]

Where $r_i$ the probability of the 'all red' alleles for the ith unique
sampling pattern (unique numbers of alleles sampled from all
populations), $g_i$ is the 'all green' equivalent, and $n_i$ is the
number of sites with the ith unique sampling numbers.

As a result, the new approach uses the correct correction when there is
heterogeneity in the number of sampled alleles among sites (i.e.,
missing data).

All tests are passing, but the new behavior is only tested in the sense
that it is not the same as the old behavior. SNAPP uses the old,
incorrect correction, so I cannot compare likelihoods to another
implementation.  Currently running simulations to vet performance of new
correction.