Updating spreadsheet module.
Adding attribute to Spreadsheet class to store the header as a vector,
so that when 'get_keys' is called, the column headers are returned in
their original order.

Updating 'parse' functions accordingly to store the header as a vector.

Updating tests accordingly; all spreadsheet tests are passing (still
need to run full test suite).