Fixing bug in 'sumcoevolity' tool.
The indices used for divergnece models output 'sumcoevolity' did not
match the order they were provided in the config. While the model
posterior and prior probabilities were correct, it was not possible to
determine which indices referred to which comparisons. This commit fixes
this problem by having the indices match the order of the comparisons in
the config file.