Cleaning up Hastings ratio code for some operators
Cleaning up the Hastings ratio logic for the TimeSizeMixer and
TimeSizeRateMixer operators. These operators had logic of the form:

if (some conditions are met) {
    hastings ratio = m^{a-b}
} else {
    hastings ratio = m^{a-b-2}
}

where 'm' is a multiplier, 'a' is the number of parameters that were
multplied by 'm', and 'b' is the number of parameters that were
multiplied by '1/m'.

The conditions were always met, and the else clause was never used.
Indeed, given that m ~ e^u, where u ~ Uniform[-1, 1], the "-2" in the
else clause is never needed.

These changes do not affect any previous results, because the else
clause was never used.

All tests of the TimeSizeMixer and TimeSizeRateMixer operators are
passing.