Debugging new model prior behavior.
There were some bugs in the logic of
BaseComparisonPopulationTreeCollection::compute_log_likelihood_and_prior
and
BaseComparisonPopulationTreeCollection::draw_heights_from_prior
due to the new handling of which model prior was being used via
EcoevolityOptions::ModelPrior. This commit fixes those.

Also, fixing the ReversibleJumpSampler operator now that the prior of
the divergence scenario is incorporated into the overall model prior by
default. This prior is incorporated into the RJ move itself, and so
computing the overall model prior with the divergence scenario included
broke this operator. Adding a boolean attribute to the operator class,
and a second boolean argument to
BaseComparisonPopulationTreeCollection::compute_log_likelihood_and_prior
so that some CollectionOperatorInterface operators can skip the
divergence scenario when calculating the overall prior density.

All test are passing!!