Updating how ReversibleJumpSampler restores state.
Updating how the ReversibleJumpSampler operator stores and restores the
state of the model. It appears that it may not have been restoring the
likelihood and prior of the trees it owns, which may be why the move was
behaving so oddly with data.

Also, cleaning up code in RandomNumberGenerator.random_set_partition.

ALL tests are passing!