Working on ReversibleJumpSampler.
Removing EventTimeScaler attribute. ReversibleJumpSampler now used
'get_time_operators()' collection method to update times rather than
it's own time operator (just like the Dirichlet and Pitman-Yor process
Gibbs samplers).

Also, making the default behavior to use a mix of jump-to-gap (80%) and
jump-to-prior (20%) moves.

Simulation tests confirmed that the last commit (the one that updated
the store/restore behavior of ReversibleJumpSampler) fixed the bug that
was preventing the uniform prior from being used. Finally!  So, this
commit also removes the warning message associated with the uniform
prior (saying it was buggy and experimental and not to use it).

All tests are passing!