Working on NAN likelihood bug.
Changing behavior of PopulationTree. Rather than throw and error if a
variable characters have a prob density of zero, it now returns -inf log
likelihood and lets the MCMC reject.

Also, checking initial likelihood in collection MCMC. If it is -inf of
NAN an error is thrown.

Updating tests of tree classes to handle new behavior of returning -inf
rather than NAN.