Starting to update sims to output locus info.
This requires keeping track of the order of site patterns and the end of
loci as sites are added to simulated BiallelicData.
Updating BiallelicData and dataset simulation methods of PopulationTree
to do this.

Still need to add method to BiallelicData to write out nexus file with
site patterns in order of linkage and charsets to delimite loci.