Reworking base of tree class hierarchy.
Creating a BaseTree class from which the population tree classes will
inherit from. This will allow the MCMC moves to explore the general
space of trees to work on the BaseTree. This will make it easier to
couple these moves with other data types/likelihood functions.

All tests are passing, so it doesn't appear that I broke anything.

The BaseTree can now keep an ordered vector of pointers to the node
height parameters in the tree (ordered by age). Also, adding
store/restore methods for this attribute. This new behavior needs more
tests.