Working on tree class hierarchy.
Adding a sensible prior on node heights to the BaseTree class.  In doing
so, I found several methods belonging to BasePopulationTree that should
either belong to BaseTree or the inherited class PopulationTree.  Moving
these methods accordingly.  Also, adding a note to change the name of
the PopulationTree class, as the name of this former base of the tree
class hierarchy is now confusing.