Working on store/restore of BaseTree.
Adding storing and restoring of the tree topology.  This appears to be
working and the store/restore test is now passing. However, I am not
sure why it is passing. It seems the current way of testing the restored
tree should fail, because the old and new node pointers shouldn't eval
to equal.

This needs to be tested further.