Updating store/restore test of BaseTree.
Previously, this was passing, but many assertions were passing for the
wrong reasons. The root node passed into the BaseTree constructor was no
longer pointing to the same root node as the BaseTree instance after the
first restore call. Because many of the assertions were performed on the
original root pointer, which was no longer being updated, it looked like
restore was working. Overhauling this test to more properly test that
the store and restore methods are working. This test is still passing.