Working on tests of SplitLumpNodesRevJumpSampler.
Working on tests of 4-leaved trees for SplitLumpNodesRevJumpSampler.
All tests of propose_merge and propose_split methods are passing,
which means that the operator is returning the Hastings Ratios
that I expect for all of the moves performed with with a tree
with 4-leaves.

However, the tests that try to sample from the prior for a tree with 4
leaves are still failing. It should be sampling all possible toplogies
at equal frequencies. But, what it is actually doing is sampling all
trees at equal frequency EXCEPT the fully resolved ladderized trees. For
some reason, it is sampling these ladderized trees exactly half as often
as all the other tree topologies. This is true for all 12 ladderized
topologies (they are sampled at equal frequency, but half as much as all
the other topolgies).

So, this problem does not seem to be a coding bug, given that all the
propose_merge/propose_split tests are returning what I think are the
correct Hastings Ratios. Perhaps my math is wrong for the Hastings
ratios.