Fixing problem with prior on node heights.
I've been having the issue for a while where when trying to sample from
the prior of a 4-tipped tree, I was sampling bifurcating ladderized
trees 1/2 as often as all other topologies.

I have been trying to troubleshoot the hastings ratio of the
SplitLumpNodesRevJumpSampler to fix this. But, Mark Holder pointed out
that this behavior was not due to the Hastings ratio, but rather the
incoherence of the prior on node heights.

I was using a uniform(0, root_height) prior on all non-root internal
node heights, but this is incoherent for nested nodes.
Now, I am using a conditional uniform(0, youngest parent of node mapped
to height) prior, which avoids this problema.
This distribution can be improved, but fow now it fixeds the problem.

All tests of SplitLumpNodesRevJumpSampler are passing!!!

To do:
-   Add tests of BaseTree::get_height_index_of_youngest_parent
-   Update tests of height operators now that the prior on heights has
    changed.