Updating BaseNode methods.
Adding BaseNode::get_deep_copy() method for getting a full, deep copy of
the node and all descendants. Currently this is untested, but also is
not used.

Updating BaseNode::get_node method. The previous test for this was
too trivial, and larger tests revealed this method wasn't working.
These updates fix this issue.