Fixing way BaseTree restores parameters.
Previously, I was reconstituting the vector of node ages too early
(prior to restoring the heights in the tree); this was leading to states
where the ages were out of order in the vector.  Now, the the node age
vector is updated as the last step of BaseTree::restore_parameters.