Reworking node indices.
Making index_ an attribute of BaseNode, which replaces the
population_index_ attribute of PopulationNode. Constructors
and methods modified accordingly.

Also, making the index an int (rather than unsigned) that is negative by
default. If get_index() is called an assertion assures the index is
positive before returning it.

Similarly, making population_index_ attribute of GeneTreeSimNode an int
that is negative by default (with assertion to ensure not negative when
get_population_index() is called.

Everything is building and tests in test_tree.cpp are passing.