Fixing bugs revealed by 5-leaf rev-jump test.
Fixing bugs in BaseNode and BaseTree classes revealed by the 5-leaf test
of the reversible-jump general tree operator.

BaseNode::split_children_from_polytomy method was missing an else
clause, which could result in the newly created node being added as a
child to multiple nodes.

Other bugs had to do with making sure nodes created during RJ move had
their split resized.

The 5-leaf test will run now, but currently our way of comparing trees
(as a set of splits and their node height indices) is faulty.  Including
the node height indices means that the same topologies will be
considered different when their node height ordering is different, which
is not what we want. We need another way of representing and comparing
topologies with shared node heights that does not restrict the ordering
of the node heights.