Working on debugging RJ topo move.
Fixed logic for keeping track of how many merged nodes result after the
propose_merge move of SplitLumpNodesRevJumpSampler.  Previously, I was
counting all nodes to be merged, but multiple nodes can end up as one
merged node, which needs to be accounted for. The same is true for
keeping track of the sizes of polytomies that are created. Previously, I
would have double counted nodes into which multiple child nodes were
merged. This should be fixed.

Also, fixing the calculation of the Bell number term of the Hasting's
ratio of the propose_split method of SplitLumpNodesRevJumpSampler. I
started the product of Bell numbers as zero, which of course, will
always be zero as terms are multiplied to it. Changing this to start as
one.

Adding additional tests for problem cases with a 5-leaf tree, which are
passing. Still need to run all the tests of
SplitLumpNodesRevJumpSampler.