SplitLumpNodesRevJumpSampler seems to be working!
In BaseTree::merge_node_height_up, I was not keeping track of the sizes
of the polytomies that were moved up by merge (but not created/made
larger by the merge). These missing polytomy sizes were borking the
Hasting's ratio calculation that depends on them.

The general 5-leaf test seems to be sampling all 336 tree models
uniformly (according to a chi-square test and eye-balling results).

Currently running the 6-leaf test that is added in this commit.