Trying more thorough 6-leaf test.
Results of first test were ambiguous. The p-value of a chi-square was
quite small (0.0113), but the number of samples across trees seemed
quite uniform by eye, and there were no obvious trends in the types of
trees that were sampled the most or least. There is a huge number of
possible trees (at least 5627), so the chi-square test is likely
dubious.