Adding checks of initial values withing priors.
Adding is_within_support method to distribution classes in
probability.hpp. This is used when ln_pdf/relative_ln_pdf is called to
see if the value is outside the support of the distribution (in which
case -inf is returned). Separating this check out as a separate method
will allow parameter and settings classes to use it to validate
arguments.

Updating PositiveRealParameterSettings to check if the initial value
specified falls within the support of the specified prior.

Updating initialize_value method of parameter classes in parameter.hpp
to check if the initial value specified falls within the support of the
specified prior.

This creates redundant checks that initial value arguments make sense in
light of the prior arguments.

All tests in test_probability.cpp, test_parameter.cpp, and
test_settings.cpp are passing.