Updating settings.
BaseCollectionSettings was setting weight of concentration scaler
to zero when the model prior was uniform. Changing this so that
split_weight can be estimated.

Also, before parsing the Dirichlet process prior settings (with
parse_dirichlet_process_prior), setting discount to 0 and fixed.

These changes required 4 tests to be updated.

All tests in test_setting.cpp are passing.