Updating probability classes.
Adding static functions to BetaDistribution so that an instance is not
necessary to get the prob density, scaled prob density, or draw a value.
The non-static funtions that did these things now use the new static
functions.

Making ContinuousProbabilityDistribution::ln_gamma_function static,
because this was necessary to make some of the BetaDistribution
functions static.

All tests of the probability classes are passing.