Updating BaseTree node height prior.
Changing prior on (non-root) node heights from uniform from
0 to youngest parent height, to beta distributed from 0 to
youngest parent. This involved adding the alpha and beta
parameters of this prior as members of the BaseTree class.

These alpha and beta parameters are currently fixed to 1.0,
so it is still equivalent to the previous uniform prior.

Full test suite was not run, but I ran enough tests to convince myself
that the changes didn't change anything; previous tests of the uniform
prior are passing under the new beta(1,1) prior.

Now I need to add functions to manipulate the alpha and beta member
parameters, and operators to update them.