Working on new BaseTree node height prior.
Changing naming scheme of the alpha and beta parameters that control the
beta prior on (non-root) node heights (these are data members of the
BaseTree class.

Adding the calculation of the prior density of alpha and beta
parameters in the log prior calculation of BaseTree.

Adding getters/setters for the prior distribution of the alpha
and beta parameters of the beta prior on node heights.

Debugging NodeHeightPriorAlphaScaler and NodeHeightPriorAlphaMover
and addoing NodeHeightPriorBetaScaler and NodeHeightPriorBetaMover.

Adding test of NodeHeightSlideBumpScaler to confirm that the beta
distribution on node heights is working as expected. This test is
passing.

Adding tests of new operators for the alpha and beta parameters.  These
are all passing.