Adding function to initialize inserted nodes.
Adding BaseNode function finish_initializing_inserted_internal_node.
This function is intended to be overridden by derived classes that have
additional priors/parameters that need to be set up when a new node is
inserted.

This primarily applies to the SplitLumpRevJumpSampler move that samples
the general tree space. This was designed to operate on the BaseTree
class, but was causing an error when trying to operate on
BasePopulationTree, because the newly created/inserted PopulationNodes
didn't have a prior on the population size.

PopulationNode now overrides this new function to solve this bug.