Adding test of a new operator.
Adding a test for one of the new operators of BasePopulationTree;
MuRateScaler.

Adding this test revealed a lot of errors in the new general
tree operator classes that operate on BasePopulationTree. Fixing
those here such that everything compiles.

The test of MuRateScaler is passing, but tests of the other new
operators are still needed.