Confirming this node height prior doesn't work.
Treating the node heights a relative does not appear coherent.  When
trying priors on this space, we cannot sample uniformly across
generalized tree topologies.

This is true for two different relative height priors:

1.  Each (non-root) height is ~ uniform(0, 1), where zero is the present
    (tips) and 1 is the youngest height with a parent node.
2.  Each (non-root) height is ~ uniform(0, 1), where zero is the oldest
    height with a child node and one is the youngest height with a
    parent node.

Abandoning this branch, but leaving it here in case we ever want to try
a Dirichlet distribution on node heights. I am pretty sure the node
sharing of the generalized tree space will cause problems for a Dirchlet
dist, so not exploring now.